+++ src/editor.js
@@ -11,7 +11,7 @@ function getContentCSS() {
         .x-todo-box {position: relative; left: -24px;}
         .x-todo-box input{position: absolute;}
         blockquote{border-left: 6px solid #ddd;padding: 5px 0 5px 10px;margin: 15px 0 15px 15px;}
-        hr{display: block;height: 0px; border: 0;border-top: 1px solid #ccc; margin: 15px 0; padding: 0;}
+        hr{display: block;height: 0; border: 0;border-top: 1px solid #ccc; margin: 15px 0; padding: 0;}
         pre{padding: 10px 5px 10px 10px;margin: 15px 0;display: block;line-height: 18px;background: #F0F0F0;border-radius: 3px;font-size: 13px; font-family: 'monaco', 'Consolas', "Liberation Mono", Courier, monospace; white-space: pre; word-wrap: normal;overflow-x: auto;}
     </style>
     `;
@@ -21,6 +21,7 @@ function createHTML(options = {}) {
     const {
         backgroundColor = '#FFF',
         color = '#000033',
+        caretColor = '',
         placeholderColor = '#a9a9a9',
         contentCSSText = '',
         cssText = '',
@@ -32,18 +33,20 @@ function createHTML(options = {}) {
         defaultParagraphSeparator = 'div',
         // When first gaining focus, the cursor moves to the end of the text
         firstFocusEnd = true,
+        useContainer = true,
     } = options;
     //ERROR: HTML height not 100%;
     return `
 <!DOCTYPE html>
 <html>
 <head>
+    <title>RN Rich Text Editor</title>
     <meta name="viewport" content="user-scalable=1.0,initial-scale=1.0,minimum-scale=1.0,maximum-scale=1.0">
     <style>
         * {outline: 0px solid transparent;-webkit-tap-highlight-color: rgba(0,0,0,0);-webkit-touch-callout: none;box-sizing: border-box;}
-        html, body { margin: 0; padding: 0;font-family: Arial, Helvetica, sans-serif; font-size:1em;}
-        body { overflow-y: hidden; -webkit-overflow-scrolling: touch;height: 100%;background-color: ${backgroundColor};}
-        .content {font-family: Arial, Helvetica, sans-serif;color: ${color}; width: 100%;height: 100%;-webkit-overflow-scrolling: touch;padding-left: 0;padding-right: 0;}
+        html, body { margin: 0; padding: 0;font-family: Arial, Helvetica, sans-serif; font-size:1em; height: 100%}
+        body { overflow-y: hidden; -webkit-overflow-scrolling: touch;background-color: ${backgroundColor};caret-color: ${caretColor};}
+        .content {font-family: Arial, Helvetica, sans-serif;color: ${color}; width: 100%;-webkit-overflow-scrolling: touch;padding-left: 0;padding-right: 0;}
         .pell { height: 100%;} .pell-content { outline: 0; overflow-y: auto;padding: 10px;height: 100%;${contentCSSText}}
     </style>
     <style>
@@ -54,7 +57,7 @@ function createHTML(options = {}) {
     <style>${cssText}</style>
 </head>
 <body>
-<div class="content"><div id="editor" class="pell"></div></div>
+<div class="content"><div id="editor" class="pell"/></div>
 <script>
     var __DEV__ = !!${window.__DEV__};
     var _ = (function (exports) {
@@ -69,7 +72,7 @@ function createHTML(options = {}) {
         var body = document.body, docEle = document.documentElement;
         var defaultParagraphSeparatorString = 'defaultParagraphSeparator';
         var formatBlock = 'formatBlock';
-        var editor = null, o_height = 0;
+        var editor = null, editorFoucs = false, o_height = 0, compositionStatus = 0, paragraphStatus = 0;
         function addEventListener(parent, type, listener) {
             return parent.addEventListener(type, listener);
         };
@@ -214,6 +217,8 @@ function createHTML(options = {}) {
         var _keyDown = false;
         function handleChange (event){
             var node = anchorNode;
+            Actions.UPDATE_HEIGHT();
+            Actions.UPDATE_OFFSET_Y();
             if (_keyDown){
                 if(_checkboxFlag === 1 && checkboxNode(node)){
                     _checkboxFlag = 0;
@@ -372,14 +377,31 @@ function createHTML(options = {}) {
             },
 
             init: function (){
-                setInterval(Actions.UPDATE_HEIGHT, 150);
-                Actions.UPDATE_HEIGHT();
+                if (${useContainer}){
+                    // setInterval(Actions.UPDATE_HEIGHT, 150);
+                    Actions.UPDATE_HEIGHT();
+                } else {
+                    body.style.height = docEle.clientHeight + 'px';
+                }
             },
 
             UPDATE_HEIGHT: function() {
-                var height = Math.max(docEle.scrollHeight, body.scrollHeight);
+                if (!${useContainer}) return;
+                // var height = Math.max(docEle.scrollHeight, body.scrollHeight);
+                var height = editor.content.scrollHeight;
                 if (o_height !== height){
                     _postMessage({type: 'OFFSET_HEIGHT', data: o_height = height});
+                }
+            },
+
+            UPDATE_OFFSET_Y: function (){
+                if (!${useContainer}) return;
+                var node = anchorNode || window.getSelection().anchorNode;
+                if (node){
+                    var offsetY = node.offsetTop || node.parentNode.offsetTop;
+                    if (offsetY){
+                        _postMessage({type: 'OFFSET_Y', data: offsetY});
+                    }
                 }
             }
         };
@@ -398,7 +420,12 @@ function createHTML(options = {}) {
             content.oninput = function (_ref) {
                 // var firstChild = _ref.target.firstChild;
                 if ((anchorNode === void 0 || anchorNode === content) && queryCommandValue(formatBlock) === ''){
-                    formatParagraph(true);
+                    if ( !compositionStatus ){
+                        formatParagraph(true);
+                        paragraphStatus = 0;
+                    } else {
+                        paragraphStatus = 1;
+                    }
                 } else if (content.innerHTML === '<br>') content.innerHTML = '';
 
                 saveSelection();
@@ -473,9 +500,14 @@ function createHTML(options = {}) {
                 ${keyDownListener} && postKeyAction(event, "CONTENT_KEYDOWN");
             }
             function handleFocus (){
+                editorFoucs = true;
+                setTimeout(function (){
+                    Actions.UPDATE_OFFSET_Y();
+                });
                 postAction({type: 'CONTENT_FOCUSED'});
             }
             function handleBlur (){
+                editorFoucs = false;
                 postAction({type: 'SELECTION_CHANGE', data: []});
                 postAction({type: 'CONTENT_BLUR'});
             }
@@ -507,6 +539,13 @@ function createHTML(options = {}) {
                     exec("insertText", text);
                 }
             });
+            addEventListener(content, 'compositionstart', function(){
+                compositionStatus = 1;
+            })
+            addEventListener(content, 'compositionend', function (){
+                compositionStatus = 0;
+                paragraphStatus && formatParagraph(true);
+            })
 
             var message = function (event){
                 var msgData = JSON.parse(event.data), action = Actions[msgData.type];